Add DataCopyScene and implement backup/restore with progress indication properly
Added a new DataCopyScene which takes care of displaying a progress bar while we're backup'ing or restoring data.

I also implemented cartridge Rom and Save backup to (micro)SD card  and Save restore from (micro)SD card
to the cartridge.

And it works! I verified it with my Pokémon Blue cartridge: I can back up its save, play the game,
release almost all pokémon, go to a different location and save again and then restore the previous save using PokeMe64

Rom backup also works: I can play the resulting rom in VisualBoyAdvance without any issue. I can also use the save
that I backed up in VisualBoyAdvance.

So right now, you can take an emulator save/pkhex save and transfer it to an actual (original) pokémon cartridge
and it will work!

However, it's not ready for release yet for a few reasons:

- It will currently always output to sd:/gb_out.sav and sd:/rom.gb. I want it to output to
sd:/<cartridgetitle>_<trainername>.sav and sd:/<cartridgetitle_<trainername>.gb instead. In fact I want PokeMe64 to
add numbers to the save file in case one already exists with that name.

- libdragon currently doesn't have the functionality to create directories. I wish PokeMe64 to backup to a directory
called PokeMe64, even if the directory doesn't exist yet.

- I want a file selector so you can select which save file to restore. This way you could have multiple saves. In
theory that would also allow you to transfer a pokémon red save to a pokémon blue cartridge and so on.

- I want to have a wipe save option

- I need to safeguard PokeMe64 for corrupted save data. Right now, for reproduction carts PokeMe64 crashes while trying
to decode the trainerName. This is because the actual save file is not accessible on a reproduction cart. If I'm going
to allow people to restore random saves (even from different versions alltogether), I need PokeMe64 to be robust enough
for people to be able to wipe the save file or restore a different one. (in order for them to be able to correct their
mistake without having to take out the battery). In order to be robust enough, I should make sure to check the CRC
checksum of the save file before going to the main menu. And if it doesn't match, I should offer a reduced menu to
the user.

So... no release of this feature until I resolve all the topics above.

But man, it's so satisfying to see it work in the current form already! The functionality itself just works on first try!

+++ include/widget/ProgressBarWidget.h
@@ -32,9 +32,9 @@ public:
     virtual ~ProgressBarWidget();
 
     /**
-     * Sets the current progress. Should be a value in the [0 - 100] interval
+     * Sets the current progress. Should be a value in the [0.0 - 1.0] interval
      */
-    void setProgress(uint8_t progress);
+    void setProgress(double progress);
 
     /**
      * Sets the ProgressBarWidget style
@@ -102,7 +102,7 @@ private:
     ProgressBarWidgetStyle style_;
     bool visible_;
     Rectangle bounds_;
-    uint8_t progress_;
+    double progress_;
     char textBuffer_[5];
 };
